Moved cache testing to its own test class
