+++ src/test/java/com/autonomouslogic/eveesiproxy/handler/ProxyHandlerTest.java
@@ -102,227 +102,4 @@ public class ProxyHandlerTest {
 
 		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
 	}
-
-	@ParameterizedTest
-	@ValueSource(strings = {"public, max-age=60, immutable"})
-	@SneakyThrows
-	void shouldCacheCachableResponses(String cacheControl) {
-		TestHttpUtils.enqueueResponse(mockEsi, 200, "Test body", Map.of("Cache-Control", cacheControl));
-
-		// First proxy response.
-		var proxyResponse1 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse1,
-				200,
-				"Test body",
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_MISS));
-
-		// ESI request.
-		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
-
-		// Second proxy response should be served from cache.
-		var proxyResponse2 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse2,
-				200,
-				"Test body",
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_HIT));
-
-		// A second request to the ESI should never be made.
-		TestHttpUtils.assertNoMoreRequests(mockEsi);
-	}
-
-	@ParameterizedTest
-	@ValueSource(strings = {"private", "no-store", "no-cache", "max-age=0"})
-	@SneakyThrows
-	void shouldNotCacheResponsesWithUncachableCacheControls(String cacheControl) {
-		for (int i = 0; i < 2; i++) {
-			TestHttpUtils.enqueueResponse(mockEsi, 200, "Test body " + i, Map.of("Cache-Control", cacheControl));
-		}
-
-		// First proxy response.
-		var proxyResponse1 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse1,
-				200,
-				"Test body 0",
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_MISS));
-
-		// ESI request.
-		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
-
-		// Second proxy response should not be served from cache.
-		var proxyResponse2 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse2,
-				200,
-				"Test body 1",
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_MISS));
-
-		// A second request to the ESI should be made.
-		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
-	}
-
-	@ParameterizedTest
-	@ValueSource(ints = {400, 404, 500, 502, 503, 504})
-	@SneakyThrows
-	void shouldNotCacheUncachableResponseCodes(int responseCode) {
-		TestHttpUtils.enqueueResponse(mockEsi, responseCode);
-		TestHttpUtils.enqueueResponse(mockEsi, 200, "Test body");
-
-		// First proxy response.
-		var proxyResponse1 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse1,
-				responseCode,
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_MISS));
-		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
-
-		// Second proxy response should not be served from cache.
-		var proxyResponse2 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse2,
-				200,
-				"Test body",
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_MISS));
-		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
-	}
-
-	@Test
-	@SneakyThrows
-	void shouldCacheResponsesWithinExpireTime() {
-		var expires = ZonedDateTime.now().plusSeconds(10);
-		var expiresString = DateTimeFormatter.RFC_1123_DATE_TIME.format(expires);
-		TestHttpUtils.enqueueResponse(
-				mockEsi, 200, "Test body", Map.of(HeaderNames.EXPIRES.lowerCase(), expiresString));
-
-		// First proxy response.
-		var proxyResponse1 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse1,
-				200,
-				"Test body",
-				Map.of(
-						ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS,
-						ProxyHeaderValues.CACHE_STATUS_MISS,
-						HeaderNames.EXPIRES.lowerCase(),
-						expiresString));
-
-		// ESI request.
-		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
-
-		// Second proxy response should be served from cache.
-		var proxyResponse2 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse2,
-				200,
-				"Test body",
-				Map.of(
-						ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS,
-						ProxyHeaderValues.CACHE_STATUS_HIT,
-						HeaderNames.EXPIRES.lowerCase(),
-						expiresString));
-
-		// A second request to the ESI should never be made.
-		TestHttpUtils.assertNoMoreRequests(mockEsi);
-	}
-
-	@ParameterizedTest
-	@ValueSource(booleans = {true, false})
-	@SneakyThrows
-	void shouldConditionallyRequestEtags(boolean modified) {
-		TestHttpUtils.enqueueResponse(mockEsi, 200, "Test body", Map.of(HeaderNames.ETAG.lowerCase(), "hash1"));
-		if (modified) {
-			TestHttpUtils.enqueueResponse(mockEsi, 200, "Test body new");
-		} else {
-			TestHttpUtils.enqueueResponse(mockEsi, 304);
-		}
-
-		// First proxy response.
-		var proxyResponse1 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse1,
-				200,
-				"Test body",
-				Map.of(
-						ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS,
-						ProxyHeaderValues.CACHE_STATUS_MISS,
-						HeaderNames.ETAG.lowerCase(),
-						"hash1"));
-
-		// ESI request.
-		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
-
-		// Second proxy response should be served from cache.
-		var proxyResponse2 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse2,
-				200,
-				modified ? "Test body new" : "Test body",
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_MISS));
-
-		// A second request to the ESI should be conditional.
-		var esiRequest = TestHttpUtils.takeRequest(mockEsi);
-		TestHttpUtils.assertRequest(esiRequest, "GET", "/esi", Map.of(HeaderNames.IF_NONE_MATCH.lowerCase(), "hash1"));
-	}
-
-	@Test
-	@SneakyThrows
-	void shouldConditionallyRequestBothExpiresAndEtags() {
-		var expires = ZonedDateTime.now().plusSeconds(2);
-		var expiresString = DateTimeFormatter.RFC_1123_DATE_TIME.format(expires);
-		TestHttpUtils.enqueueResponse(
-				mockEsi,
-				200,
-				"Test body",
-				Map.of(HeaderNames.EXPIRES.lowerCase(), expiresString, HeaderNames.ETAG.lowerCase(), "hash1"));
-
-		// First proxy response.
-		var proxyResponse1 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse1,
-				200,
-				"Test body",
-				Map.of(
-						ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS,
-						ProxyHeaderValues.CACHE_STATUS_MISS,
-						HeaderNames.EXPIRES.lowerCase(),
-						expiresString,
-						HeaderNames.ETAG.lowerCase(),
-						"hash1"));
-
-		// ESI request.
-		assertNotNull(TestHttpUtils.takeRequest(mockEsi));
-
-		// Second proxy response should be served from cache.
-		var proxyResponse2 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse2,
-				200,
-				"Test body",
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_HIT));
-
-		// No requests should have been made to the ESI.
-		TestHttpUtils.assertNoMoreRequests(mockEsi);
-
-		// Wait for the cache to expire.
-		while (ZonedDateTime.now().isBefore(expires.plusSeconds(2))) {
-			Thread.sleep(10);
-		}
-
-		// The ESI should respond with a 304 Not Modified.
-		TestHttpUtils.enqueueResponse(mockEsi, 304);
-
-		// Third proxy response should be conditional.
-		var proxyResponse3 = TestHttpUtils.callProxy(client, proxy, "GET", "/esi");
-		TestHttpUtils.assertResponse(
-				proxyResponse3,
-				200,
-				"Test body",
-				Map.of(ProxyHeaderNames.X_EVE_ESI_PROXY_CACHE_STATUS, ProxyHeaderValues.CACHE_STATUS_MISS));
-
-		// A third request to the ESI should be conditional.
-		var esiRequest = TestHttpUtils.takeRequest(mockEsi);
-		TestHttpUtils.assertRequest(esiRequest, "GET", "/esi", Map.of(HeaderNames.IF_NONE_MATCH.lowerCase(), "hash1"));
-	}
 }
